+++ youtokentome/cpp/bpe.h
@@ -1,13 +1,56 @@
 #pragma once
 
-#include <map>
 #include <string>
 #include <unordered_set>
+#include <vector>
+
 #include "third_party/flat_hash_map/flat_hash_map/flat_hash_map.h"
 
 #include "utils.h"
 
+// TODO: introduce vkcom::bpe namespace
 namespace vkcom {
+
+struct BPE_Rule {
+  // x + y -> z
+  uint32_t x{0};
+  uint32_t y{0};
+  uint32_t z{0};
+
+  BPE_Rule() = default;
+
+  BPE_Rule(uint32_t x, uint32_t y, uint32_t z);
+
+  bool operator==(const BPE_Rule &other) const;
+};
+
+struct BpeConfig {
+  double character_coverage = 1;
+  int n_threads = 0;
+  SpecialTokens special_tokens;
+
+  BpeConfig() = default;
+
+  BpeConfig(double character_coverage, int n_threads,
+            const SpecialTokens &special_tokens);
+};
+
+struct BPEState {
+  flat_hash_map<uint32_t, uint32_t> char2id;
+  std::vector<BPE_Rule> rules;
+  SpecialTokens special_tokens;
+
+  void dump(const std::string &file_name);
+
+  Status load(const std::string &file_name);
+};
+
+struct EncodingConfig {
+  bool bos;
+  bool eos;
+  bool reverse;
+  double dropout_prob;
+};
 
 Status train_bpe(const std::string &input_path, const std::string &model_path,
                  int vocab_size, BpeConfig config);
